+++ src/isobands/__init__.py
@@ -1,5 +1,5 @@
 """Create filled-contour polygons from xarray data."""
 
-from isobands.core import isobands
+from isobands.core import gdal_fixed_level_polygons, isobands
 
-__all__ = ["isobands"]
+__all__ = ["gdal_fixed_level_polygons", "isobands"]